@@ -1,51 +1,57 @@
-buildscript {
-    repositories {
-        google()
-        jcenter()
-        mavenCentral()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.2'
-    }
-}
 apply plugin: 'com.android.application'
 
-
 android {
-    compileSdkVersion 27
-    buildToolsVersion "28.0.3"
-
+    compileSdkVersion 34
+    buildToolsVersion "34.0.0"
+    namespace "com.tsymiar.SerialConn"
     defaultConfig {
         applicationId "com.tsymiar.SerialConn"
-        minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 7
+        minSdkVersion 21
+        //noinspection EditedTargetSdkVersion
+        targetSdkVersion 33
+        versionCode 8
         versionName "3.3.3"
     }
     buildTypes {
         release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            minifyEnabled true
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
         debug {
             debuggable true
-            jniDebuggable true
             renderscriptDebuggable true
             minifyEnabled true
             pseudoLocalesEnabled true
             zipAlignEnabled true
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_21
+        targetCompatibility JavaVersion.VERSION_21
+    }
+    testOptions {
+        unitTests.all {
+            enabled = false
+        }
+    }
 }
 
 dependencies {
     //noinspection GradleCompatible
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:support-annotations:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.7.0'
+    implementation 'com.google.android.material:material:1.12.0'
+    implementation 'androidx.annotation:annotation:1.9.1'
+    //noinspection GradleDependency
+    implementation 'androidx.core:core-ktx:1.12.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.2.1'
     implementation fileTree(dir: 'libs', include: ['*.aar', '*.jar'], exclude: [])
-    implementation files('libs\\achartengine-1.1.0.jar')
+    implementation files('libs/achartengine-1.1.0.jar')
 }
 
-repositories {
-    google()
+allprojects {
+    gradle.projectsEvaluated {
+        tasks.withType(JavaCompile) {
+            // options.compilerArgs << "-Xlint:deprecation"
+        }
+    }
 }
