@@ -2,7 +2,7 @@ package com.tsymiar.SerialConn;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.KeyEvent;
 import android.widget.Toast;
 
